refactor: extract segment groups to new files
