+++ assets/wizards/popups/components/segment-group/icons.js
@@ -1,0 +1,35 @@
+/**
+ * WordPress dependencies.
+ */
+import { Path, SVG } from '@wordpress/components';
+
+export const iconInline = (
+	<SVG xmlns="http://www.w3.org/2000/svg" height="24" viewBox="0 0 24 24" width="24">
+		<Path d="M21 11.01L3 11v2h18zM3 16h12v2H3zM21 6H3v2.01L21 8z" />
+	</SVG>
+);
+export const iconCenterOverlay = (
+	<SVG xmlns="http://www.w3.org/2000/svg" height="24" viewBox="0 0 24 24" width="24">
+		<Path d="M21 3H3c-1.1 0-2 .9-2 2v14c0 1.1.9 2 2 2h18c1.1 0 2-.9 2-2V5c0-1.1-.9-2-2-2zm0 16H3V5h18v14zM7 9h10v6H7V9z" />
+	</SVG>
+);
+export const iconTopOverlay = (
+	<SVG xmlns="http://www.w3.org/2000/svg" height="24" viewBox="0 0 24 24" width="24">
+		<Path d="M3 21h18c1.1 0 2-.9 2-2V5c0-1.1-.9-2-2-2H3c-1.1 0-2 .9-2 2v14c0 1.1.9 2 2 2zM3 5h18v14H3V5zm16 4H5V6h14v3z" />
+	</SVG>
+);
+export const iconBottomOverlay = (
+	<SVG xmlns="http://www.w3.org/2000/svg" height="24" viewBox="0 0 24 24" width="24">
+		<Path d="M21 3H3c-1.1 0-2 .9-2 2v14c0 1.1.9 2 2 2h18c1.1 0 2-.9 2-2V5c0-1.1-.9-2-2-2zm0 16H3V5h18v14zM5 15h14v3H5z" />
+	</SVG>
+);
+export const iconAboveHeader = (
+	<SVG xmlns="http://www.w3.org/2000/svg" height="24" viewBox="0 0 24 24" width="24">
+		<Path d="M3 21h18c1.1 0 2-.9 2-2V5c0-1.1-.9-2-2-2H3c-1.1 0-2 .9-2 2v14c0 1.1.9 2 2 2zM3 8h18v11H3V8z" />
+	</SVG>
+);
+export const iconManualPlacement = (
+	<SVG xmlns="http://www.w3.org/2000/svg" height="24" viewBox="0 0 24 24" width="24">
+		<Path d="M21 3H3c-1.1 0-2 .9-2 2v14c0 1.1.9 2 2 2h18c1.1 0 2-.9 2-2V5c0-1.1-.9-2-2-2zM3 8h11.5v4.5H3V8zm0 6.5h11.5V19H3v-4.5zM21 19h-4.5V8H21v11z" />
+	</SVG>
+);
