+++ assets/wizards/popups/views/popup-group/index.js
@@ -6,7 +6,7 @@
  * WordPress dependencies.
  */
 import { useEffect, useState, Fragment } from '@wordpress/element';
-import { MenuItem, Path, SVG } from '@wordpress/components';
+import { MenuItem } from '@wordpress/components';
 import { __ } from '@wordpress/i18n';
 import { ENTER, ESCAPE } from '@wordpress/keycodes';
 import { Icon, moreVertical, close } from '@wordpress/icons';
@@ -30,206 +30,11 @@ import {
 	TextControl,
 	ToggleControl,
 } from '../../../../components/src';
-import PopupActionCard from '../../components/popup-action-card';
-import SegmentationPreview from '../../components/segmentation-preview';
-import { filterOutUncategorized, isOverlay, frequencyForPopup } from '../../utils';
+import SegmentGroup from '../../components/segment-group';
 import './style.scss';
 
 const { useParams } = Router;
 
-const descriptionForPopup = ( prompt, segments ) => {
-	const {
-		categories,
-		campaign_groups: campaigns,
-		sitewide_default: sitewideDefault,
-		options,
-		status,
-	} = prompt;
-	const segment = find( segments, [ 'id', options.selected_segment_id ] );
-	const filteredCategories = filterOutUncategorized( categories );
-	const descriptionMessages = [];
-	if ( campaigns.length > 0 ) {
-		const campaignsList = campaigns.map( ( { name } ) => name ).join( ', ' );
-		descriptionMessages.push(
-			( campaigns.length === 1
-				? __( 'Campaign: ', 'newspack' )
-				: __( 'Campaigns: ', 'newspack' ) ) + campaignsList
-		);
-	}
-	if ( sitewideDefault ) {
-		descriptionMessages.push( __( 'Sitewide default', 'newspack' ) );
-	}
-	if ( filteredCategories.length > 0 ) {
-		descriptionMessages.push(
-			__( 'Categories: ', 'newspack' ) +
-				filteredCategories.map( category => category.name ).join( ', ' )
-		);
-	}
-	if ( 'pending' === status ) {
-		descriptionMessages.push( __( 'Pending review', 'newspack' ) );
-	}
-	if ( 'future' === status ) {
-		descriptionMessages.push( __( 'Scheduled', 'newspack' ) );
-	}
-	descriptionMessages.push( __( 'Frequency: ', 'newspack' ) + frequencyForPopup( prompt ) );
-	return descriptionMessages.length ? descriptionMessages.join( ' | ' ) : null;
-};
-
-const getCardClassName = ( { options, sitewide_default: sitewideDefault, status } ) => {
-	if ( 'draft' === status || 'pending' === status || 'future' === status ) {
-		return 'newspack-card__is-disabled';
-	}
-	if ( sitewideDefault ) {
-		return 'newspack-card__is-primary';
-	}
-	if ( isOverlay( { options } ) && ! sitewideDefault ) {
-		return 'newspack-card__is-disabled';
-	}
-	return 'newspack-card__is-supported';
-};
-
-const iconInline = (
-	<SVG xmlns="http://www.w3.org/2000/svg" height="24" viewBox="0 0 24 24" width="24">
-		<Path d="M21 11.01L3 11v2h18zM3 16h12v2H3zM21 6H3v2.01L21 8z" />
-	</SVG>
-);
-const iconCenterOverlay = (
-	<SVG xmlns="http://www.w3.org/2000/svg" height="24" viewBox="0 0 24 24" width="24">
-		<Path d="M21 3H3c-1.1 0-2 .9-2 2v14c0 1.1.9 2 2 2h18c1.1 0 2-.9 2-2V5c0-1.1-.9-2-2-2zm0 16H3V5h18v14zM7 9h10v6H7V9z" />
-	</SVG>
-);
-const iconTopOverlay = (
-	<SVG xmlns="http://www.w3.org/2000/svg" height="24" viewBox="0 0 24 24" width="24">
-		<Path d="M3 21h18c1.1 0 2-.9 2-2V5c0-1.1-.9-2-2-2H3c-1.1 0-2 .9-2 2v14c0 1.1.9 2 2 2zM3 5h18v14H3V5zm16 4H5V6h14v3z" />
-	</SVG>
-);
-const iconBottomOverlay = (
-	<SVG xmlns="http://www.w3.org/2000/svg" height="24" viewBox="0 0 24 24" width="24">
-		<Path d="M21 3H3c-1.1 0-2 .9-2 2v14c0 1.1.9 2 2 2h18c1.1 0 2-.9 2-2V5c0-1.1-.9-2-2-2zm0 16H3V5h18v14zM5 15h14v3H5z" />
-	</SVG>
-);
-const iconAboveHeader = (
-	<SVG xmlns="http://www.w3.org/2000/svg" height="24" viewBox="0 0 24 24" width="24">
-		<Path d="M3 21h18c1.1 0 2-.9 2-2V5c0-1.1-.9-2-2-2H3c-1.1 0-2 .9-2 2v14c0 1.1.9 2 2 2zM3 8h18v11H3V8z" />
-	</SVG>
-);
-const iconManualPlacement = (
-	<SVG xmlns="http://www.w3.org/2000/svg" height="24" viewBox="0 0 24 24" width="24">
-		<Path d="M21 3H3c-1.1 0-2 .9-2 2v14c0 1.1.9 2 2 2h18c1.1 0 2-.9 2-2V5c0-1.1-.9-2-2-2zM3 8h11.5v4.5H3V8zm0 6.5h11.5V19H3v-4.5zM21 19h-4.5V8H21v11z" />
-	</SVG>
-);
-
-const Segment = ( {
-	segment,
-	campaignGroup,
-	deletePopup,
-	previewPopup,
-	segments,
-	setTermsForPopup,
-	setSitewideDefaultPopup,
-	updatePopup,
-	publishPopup,
-	unpublishPopup,
-} ) => {
-	const [ addNewPopoverIsVisible, setAddNewPopoverIsVisible ] = useState();
-	const { label, id, items } = segment;
-	return (
-		<Card isSmall className="newspack-campaigns__popup-group__card">
-			<h3 className="newspack-campaigns__popup-group__card__segment">
-				{ __( 'Segment: ', 'newspack' ) }
-				{ label }
-				<SegmentationPreview
-					campaignGroups={ [ campaignGroup ] }
-					segment={ id }
-					showUnpublished={ true } // Do we need a control for this?
-					renderButton={ ( { showPreview } ) => (
-						<Button isTertiary isSmall isLink onClick={ () => showPreview() }>
-							{ __( 'Preview', 'newspack' ) }
-						</Button>
-					) }
-				/>
-			</h3>
-			<Card noBorder className="newspack-campaigns__popup-group__action-cards">
-				{ items.map( item => (
-					<PopupActionCard
-						className={ getCardClassName( item ) }
-						deletePopup={ deletePopup }
-						description={ descriptionForPopup( item, segments ) }
-						key={ item.id }
-						popup={ item }
-						previewPopup={ previewPopup }
-						segments={ segments }
-						setTermsForPopup={ setTermsForPopup }
-						setSitewideDefaultPopup={ setSitewideDefaultPopup }
-						updatePopup={ updatePopup }
-						publishPopup={ publishPopup }
-						unpublishPopup={ unpublishPopup }
-					/>
-				) ) }
-			</Card>
-			{ items.length < 1 ? <p>{ __( 'No prompts in this segment yet.', 'newspack' ) }</p> : '' }
-			{ parseInt( campaignGroup ) > 0 && (
-				<div className="newspack-campaigns__popup-group__add-new-wrap">
-					<Button
-						isSmall
-						isTertiary
-						onClick={ () => setAddNewPopoverIsVisible( ! addNewPopoverIsVisible ) }
-					>
-						{ __( 'Add New Prompt', 'newspack' ) }
-					</Button>
-					{ addNewPopoverIsVisible && (
-						<Modal
-							title={ __( 'Add New Prompt', 'newspack' ) }
-							className="newspack-campaigns__popup-group__add-new-button__modal"
-							onRequestClose={ () => setAddNewPopoverIsVisible( false ) }
-							shouldCloseOnEsc={ false }
-							shouldCloseOnClickOutside={ false }
-						>
-							<Card buttonsCard noBorder className="newspack-card__buttons-prompt">
-								<Button
-									href={ `/wp-admin/post-new.php?post_type=newspack_popups_cpt&group=${ campaignGroup }&segment=${ id }` }
-								>
-									<Icon icon={ iconInline } />
-									{ __( 'Inline', 'newspack' ) }
-								</Button>
-								<Button
-									href={ `/wp-admin/post-new.php?post_type=newspack_popups_cpt&placement=overlay-center&group=${ campaignGroup }&segment=${ id }` }
-								>
-									<Icon icon={ iconCenterOverlay } />
-									{ __( 'Center Overlay', 'newspack' ) }
-								</Button>
-								<Button
-									href={ `/wp-admin/post-new.php?post_type=newspack_popups_cpt&placement=overlay-top&group=${ campaignGroup }&segment=${ id }` }
-								>
-									<Icon icon={ iconTopOverlay } />
-									{ __( 'Top Overlay', 'newspack' ) }
-								</Button>
-								<Button
-									href={ `/wp-admin/post-new.php?post_type=newspack_popups_cpt&placement=overlay-bottom&group=${ campaignGroup }&segment=${ id }` }
-								>
-									<Icon icon={ iconBottomOverlay } />
-									{ __( 'Bottom Overlay', 'newspack' ) }
-								</Button>
-								<Button
-									href={ `/wp-admin/post-new.php?post_type=newspack_popups_cpt&placement=above-header&group=${ campaignGroup }&segment=${ id }` }
-								>
-									<Icon icon={ iconAboveHeader } />
-									{ __( 'Above Header', 'newspack' ) }
-								</Button>
-								<Button
-									href={ `/wp-admin/post-new.php?post_type=newspack_popups_cpt&placement=manual&group=${ campaignGroup }&segment=${ id }` }
-								>
-									<Icon icon={ iconManualPlacement } />
-									{ __( 'Manual Placement', 'newspack' ) }
-								</Button>
-							</Card>
-						</Modal>
-					) }
-				</div>
-			) }
-		</Card>
-	);
-};
 /**
  * Popup group screen
  */
@@ -632,7 +437,7 @@ const PopupGroup = ( {
 				</div>
 			</div>
 			{ campaignsToDisplay.map( ( segment, index ) => (
-				<Segment
+				<SegmentGroup
 					key={ index }
 					segment={ segment }
 					campaignGroup={ campaignGroup }
